modules/locationConsumer/pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Josh Haines", email = "Josh@JoshHaines.com"},
]
dependencies = [
    "kafka-python>=2.0.2",
    "psycopg2-binary>=2.9.3",
]
requires-python = ">=3.9"
license = {text = "MIT"}

[project.urls]
homepage = ""

[tool]
[tool.pdm]

[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"
